# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parameter E-commerece API",
        "description": "Documentation for apis that can be used and their responses",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "index",
                "operationId": "1ce27f90dd8c39cea9520bf45d58c7c2",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter with trending or new-arrival"
                    }
                ],
                "requestBody": {
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Products retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products/search": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Search",
                "operationId": "e8ceb2cfd0011db14b0d8b0e59c537b2",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Search by title"
                    }
                ],
                "requestBody": {
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Products retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{category_id}/products": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Store",
                "operationId": "adf9c2ceebed528d010eccc067acb8b5",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "stock_quantity",
                                    "price",
                                    "currency",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Write the title of the product",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "stock_quantity": {
                                        "description": "Write the quantity in stock",
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "price": {
                                        "description": "Write the price of the product",
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "currency": {
                                        "description": "Write the currency name",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "trending": {
                                        "description": "Is the product trend   ing?",
                                        "type": "boolean",
                                        "enum": [
                                            1,
                                            0
                                        ]
                                    },
                                    "description": {
                                        "description": "Write the description of the product",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "file"
                                        },
                                        "collectionFormat": "multi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "data": {
                                            "properties": {
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The title already exists"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/products/{product}/request": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Product Request",
                "operationId": "dc1ca83964d386ead870135751b26d99",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product requested successfully"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}